`tools/generator-go-sdk`: refactoring to use the same CLI setup as elsewhere for consistency
This also plumbs through the same logging setup for the same reason - but doesn't refactor over to using it.
